HQR related utils moved to folder. Created OpenHqr reader
